# web/references/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Basic Todo Web API",
    "version": "v1"
  },
  "paths": {
...
    "/api/TodoItems": {
      "get": {
        "tags": [
          "TodoItems"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TodoItemResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TodoItemResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TodoItemResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "TodoItems"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TodoItemDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TodoItemDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/TodoItemDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TodoItemResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TodoItemResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TodoItemResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/TodoItems/{id}": {
      "get": {
        "tags": [
          "TodoItems"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TodoItemResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TodoItemResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TodoItemResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "TodoItems"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TodoItemDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TodoItemDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/TodoItemDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "TodoItems"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "boolean",
                "default": false
              }
            },
            "text/json": {
              "schema": {
                "type": "boolean",
                "default": false
              }
            },
            "application/*+json": {
              "schema": {
                "type": "boolean",
                "default": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TodoItemDeletionResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TodoItemDeletionResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TodoItemDeletionResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/TodoItems/bulk": {
      "post": {
        "tags": [
          "TodoItems"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/TodoItemDTO"
                }
              }
            },
            "text/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/TodoItemDTO"
                }
              }
            },
            "application/*+json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/TodoItemDTO"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TodoItemsMultipleResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TodoItemsMultipleResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TodoItemsMultipleResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "TodoItemDTO": {
        "required": [
          "title"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "isComplete": {
            "type": "boolean"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "dueDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "isDeleted": {
            "type": "boolean",
            "nullable": true
          },
          "deletedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "TodoItemDeletionResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "errorMessage": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "TodoItemResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "errorMessage": {
            "type": "string",
            "nullable": true
          },
          "data": {
            "$ref": "#/components/schemas/TodoItemDTO"
          }
        },
        "additionalProperties": false
      },
      "TodoItemsMultipleResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "errorMessage": {
            "type": "string",
            "nullable": true
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TodoItemDTO"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Cookie Authentication": {
        "type": "apiKey",
        "description": "We use cookies for authentication.",
        "name": "BasicTodo.API-Auth",
        "in": "cookie"
      }
    }
  },
  "security": [
    {
      "Cookie Authentication": [ ]
    }
  ]
}